+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 - Added Logwriter and Log
 - Added Leaf View
 - Merged the Leaf Veins Templating engine and Leaf Core
-- Added \Jenssegers\Blade\Blade to composer dependencies
+- Added Support for blade templating with Leaf Blade
 
 
 #### Fixed